+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree.rs
@@ -1,4 +1,4 @@
+pub mod create_tree;
 pub mod errors;
 pub mod node;
-pub mod original_skeleton_calc;
 pub mod tree;
